Scripts/Steam Vault & Underbog: Minor updates (#31090)
* Kalithresh now forces all distillers to cast SPELL_QUIET_SUICIDE when encounter is finished
* Move go_main_chambers_access_panel script from instance file to instance scripts file
* Add comments for script names for Hungarfen and Black Stalker
* Remove redundant SetBossState for Musel'ek (currently is handled in BossAI)

(cherry picked from commit e899b2aeea3ef2718beefb59ab0fb170cee80fc4)
