+++ src/server/scripts/Outland/CoilfangReservoir/SteamVault/instance_steam_vault.cpp
@@ -19,27 +19,8 @@
 #include "Creature.h"
 #include "CreatureAI.h"
 #include "GameObject.h"
-#include "GameObjectAI.h"
 #include "InstanceScript.h"
 #include "steam_vault.h"
-
-struct go_main_chambers_access_panel : public GameObjectAI
-{
-    go_main_chambers_access_panel(GameObject* go) : GameObjectAI(go), _instance(go->GetInstanceScript()) { }
-
-    bool OnGossipHello(Player* /*player*/) override
-    {
-        if (Creature* controller = _instance->GetCreature(DATA_DOOR_CONTROLLER))
-            controller->AI()->Talk(CONTROLLER_TEXT_ACESS_USED);
-        _instance->SetData(ACTION_OPEN_DOOR, 0);
-        me->SetFlag(GO_FLAG_NOT_SELECTABLE);
-        me->SetGoState(GO_STATE_ACTIVE);
-        return true;
-    }
-
-private:
-    InstanceScript* _instance;
-};
 
 static constexpr ObjectData gameObjectData[] =
 {
@@ -139,6 +120,5 @@ class instance_steam_vault : public InstanceMapScript
 
 void AddSC_instance_steam_vault()
 {
-    RegisterGameObjectAI(go_main_chambers_access_panel);
     new instance_steam_vault();
 }
